requirements.txt
scikit-learn
pandas
python-dotenv
matplotlib 
seaborn